fix: make `ffpopsim_hihgd.all_polymorphic` public and read-write
This property is set in some of the python examples, however it is declared and behaves as read-only and in fact is private in C++. I am not sure how it worked before, with swig 2. It might have been silently failing. Now, with swig 3+ it throws an error. 

In order to fix the error, I make this property read-write and make it public.